Fix probe
